removed 'pages' folder, moved components that were previously inside it to components folder
