+++ src/App.js
@@ -1,6 +1,6 @@
 import './App.scss';
 import Navbar from './components/Navbar';
-import Home from './pages/Home';
+import Home from './components/Home';
 
 function App() {
   return (
