+++ src/components/Home.scss
